pyproject.toml
[project]
name = "ThoraxAI"
version = "1.0.0"
description = "Classification d'images medicales -- Detection de pneumonie sur radiographies thoraciques (Projet B3 Deep Learning, ECE Paris)"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    # Deep Learning
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    # Data & Visualization
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    # ML metrics
    "scikit-learn>=1.3.0",
    # Progress bars
    "tqdm>=4.65.0",
    # Image processing
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    # Notebooks
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

[project.optional-dependencies]
demo = [
    "streamlit>=1.28.0",
]
kaggle = [
    "kaggle>=1.5.0",
]

[dependency-groups]
dev = [
    "ruff>=0.4.0",
]
